builds/backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
dnspython==2.6.1
email-validator==2.1.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.3
fonttools==4.51.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
jinja2==3.1.4
jupyter-client==8.6.1
jupyter-core==5.7.2
kiwisolver==1.4.5
lazy-loader==0.4
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mne==1.7.0
nest-asyncio==1.6.0
numpy==1.26.4
orjson==3.10.3
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
pip==24.0
platformdirs==4.2.1
pooch==1.8.1
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.1
pydantic-core==2.18.2
pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pyyaml==6.0.1
pyzmq==26.0.3
requests==2.31.0
rich==13.7.1
scipy==1.13.0
shellingham==1.5.4
-e file:///Users/ernie/Documents/GitHub/signalflow_portal/builds/docker_backend/github/signalfloweeg_py
six==1.16.0
sniffio==1.3.1
sqlalchemy==2.0.30
sqlalchemy-utils==0.41.2
stack-data==0.6.3
starlette==0.37.2
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typer==0.12.3
typing-extensions==4.11.0
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
